+++ src/api/index.js
@@ -1,3 +1,4 @@
-export { default as request } from './request'
+export { default as request, abortRequest, abortAllRequests } from './request'
+export { default as upload } from './upload'
 export { TOKEN_KEY } from '../store/auth'
 export * from './user'